// src/app/core/api/models/ahb-pruefi-diff.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AhbDiffLine } from '../models/ahb-diff-line';

/**
 * Diff comparison between two different Pruefidentifikatoren within the same format version (as opposed to AhbDiff, which compares the same Pruefidentifikator across two format versions)
 */
export interface AhbPruefiDiff {
  lines: Array<AhbDiffLine>;
  meta: {
'format_version': string;
'pruefidentifikator_old': string;
'pruefidentifikator_new': string;

/**
 * Description of the new (second) Pruefidentifikator
 */
'description_new'?: string | null;

/**
 * Description of the old (first) Pruefidentifikator
 */
'description_old'?: string | null;
};
}